move from common to test_dataset + move some additional tests to test_writer
